# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Health Habit Hub API
  version: 1.0.0
  description: >-
    REST API for the Health Habit Hub platform. Manages participants, surveys, shared habits, recommendations, and
    profiles. All endpoints (except /health) require a Keycloak JWT bearer token.
  contact:
    name: HHH Development Team
  license:
    name: MIT
...
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://hhh.example.com/api/v1
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak-issued JWT. Obtain from POST /realms/hhh/protocol/openid-connect/token
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          example: Internal server error
...
    Profile:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        answers:
          type: object
          additionalProperties: true
          example:
            age: 30
            goal: exercise more
        completedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
security:
  - bearerAuth: []
paths:
...
  /profile:
    get:
      summary: Get the authenticated user's profile
      description: Returns the profile (questionnaire answers + timestamps) for the caller identified by the JWT sub claim.
      tags:
        - Profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
              example:
                userId: a1b2c3d4-1234-5678-abcd-ef0123456789
                answers:
                  age: 30
                  goal: exercise more
                completedAt: '2026-03-15T10:00:00.000Z'
                updatedAt: '2026-03-15T10:00:00.000Z'
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Unauthorized
        '404':
          description: Profile not found for this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Profile not found
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Create or update the authenticated user's profile
      description: Upserts the profile for the caller. Merges answers and updates timestamps.
      tags:
        - Profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                answers:
                  type: object
                  additionalProperties: true
                  example:
                    age: 30
                    goal: exercise more
                completedAt:
                  type: string
                  format: date-time
                  example: '2026-03-15T10:00:00.000Z'
      responses:
        '200':
          description: Profile upserted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Profile'
              example:
                userId: a1b2c3d4-1234-5678-abcd-ef0123456789
                answers:
                  age: 30
                  goal: exercise more
                completedAt: '2026-03-15T10:00:00.000Z'
                updatedAt: '2026-03-15T10:00:00.000Z'
        '401':
          description: Missing or invalid JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                error: Unauthorized
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...